Update signs in CHSH inequality equations to be consistent with rest of document.
The first equation (image1), C = <AB> - <AB'> + <A'B> + <A'B'> is inconsistent
with the choice of measurements (A=Z, A'=X, B=W, B'=V). With the ideal
correlated expectation value of -1/sqrt(2) for <XV> (<A'B'>), and 1/sqrt(2)
for the other three, C = 0, instead of 2*sqrt(2).

The updated image1 is C = <AB> + <AB'> + <A'B> - <A'B'>.

The fourth equation (image4) has the same change applied.

Note that the URLs for image1 and image4 need to be updated to begin with
https://github.com/Qiskit/iqx-user-guide/blob/master/.